+++ src/grids/hexOfCir.ts
@@ -1,5 +1,5 @@
 import { hexOfTri } from ".";
-import { GridPoints, IGeneratorArgs} from "../GridGenerator";
+import { GridPoints, IGeneratorArgs} from "./_base";
 
 export function hexOfCir(args: IGeneratorArgs): GridPoints {
     return hexOfTri(args);