+++ src/grids/rectOfRects.ts
@@ -1,4 +1,4 @@
-import { GridPoints, IGeneratorArgs, IPoint} from "../GridGenerator";
+import { GridPoints, IGeneratorArgs, IPoint} from "./_base";
 
 export function rectOfRects(args: IGeneratorArgs): GridPoints {
     let cellWidth: number = 50;
@@ -21,11 +21,20 @@ export function rectOfRects(args: IGeneratorArgs): GridPoints {
         gridWidth = args.gridWidth;
     }
 
+    let startx: number = 0;
+    let starty: number = 0;
+    if (args.startx !== undefined) {
+        startx = args.startx;
+    }
+    if (args.starty !== undefined) {
+        starty = args.starty;
+    }
+
     const grid: GridPoints = [];
     for (let row = 0; row < gridHeight; row++) {
         const node: IPoint[] = [];
         for (let col = 0; col < gridWidth; col++) {
-            const point: IPoint = {x: (cellWidth * col), y: (cellHeight * row)};
+            const point: IPoint = {x: startx + (cellWidth * col), y: starty + (cellHeight * row)};
             node.push(point);
         }
         grid.push(node);