+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Entropy board added as a special renderer. I want to minimize the number of special renderers, but there may end up being a few.
+
 ## [0.3.0] - 2021-10-21
 
 ### Added